requirements.txt
python-telegram-bot==12.7
pip~=20.2.1
cryptography~=3.4.6
setuptools~=49.2.1
future~=0.18.2
requests~=2.22.0
tornado~=6.1
Flask==1.1.2